+++ src/library/zoid/message/component.js
@@ -14,14 +14,15 @@ import {
     getSessionID,
     getGlobalState,
     getCurrentTime,
-    writeStorageID,
+    writeToLocalStorage,
     getOrCreateStorageID,
     getStageTag,
     getFeatures,
     ppDebug,
     isScriptBeingDestroyed,
     getDevTouchpoint,
-    getMerchantConfig
+    getMerchantConfig,
+    getTsCookieFromStorage
 } from '../../../utils';
 import validate from './validation';
 import containerTemplate from './containerTemplate';
@@ -132,7 +133,7 @@ export default createGlobalVariableGetter('__paypal_credit_message__', () =>
                             onApply,
                             getContainer
                         } = props;
-                        const { offerType, messageRequestId } = meta;
+                        const { offerType, offerCountry, messageRequestId } = meta;
 
                         // Avoid spreading message props because both message and modal
                         // zoid components have an onClick prop that functions differently
@@ -144,6 +145,7 @@ export default createGlobalVariableGetter('__paypal_credit_message__', () =>
                             buyerCountry,
                             onApply,
                             offer: offerType,
+                            offerCountry,
                             refId: messageRequestId,
                             refIndex: index,
                             src: 'message_click',
@@ -203,15 +205,15 @@ export default createGlobalVariableGetter('__paypal_credit_message__', () =>
                 queryParam: false,
                 value: ({ props }) => {
                     const { onReady } = props;
-                    return ({ meta, activeTags, deviceID, requestDuration }) => {
-                        const { account, merchantId, index, modal, getContainer, messageRequestId } = props;
+                    return ({ meta, activeTags, deviceID, ts, requestDuration, messageRequestId }) => {
+                        const { account, merchantId, index, modal, getContainer } = props;
                         const { trackingDetails, offerType, ppDebugId } = meta;
                         const partnerClientId = merchantId && account.slice(10); // slice is to remove the characters 'client-id:' from account name
 
                         ppDebug(`Message Correlation ID: ${ppDebugId}`);
-
-                        // Write deviceID from iframe localStorage to merchant domain localStorage
-                        writeStorageID(deviceID);
+                        const tsCookie = typeof ts !== 'undefined' ? ts : getTsCookieFromStorage();
+                        // Write deviceID and ts cookie from iframe localStorage to merchant domain localStorage
+                        writeToLocalStorage({ id: deviceID, ts: tsCookie });
 
                         logger.addMetaBuilder(existingMeta => {
                             // Remove potential existing meta info
@@ -228,6 +230,7 @@ export default createGlobalVariableGetter('__paypal_credit_message__', () =>
                                 // Need to merge global attribute here due to preserve performance attributes
                                 global: {
                                     ...existingGlobal,
+                                    ts: tsCookie,
                                     deviceID, // deviceID from internal iframe storage
                                     sessionID: getSessionID() // Session ID from parent local storage,
                                 },