+++ src/server/locale/US/index.js
@@ -1,14 +1,20 @@
-import gpl from './GPL/index';
-import ppc from './PPC/index';
+import payLaterShortTerm from './PAY_LATER_SHORT_TERM';
+import paypalCredit from './PAYPAL_CREDIT';
+import payLaterLongTerm from './PAY_LATER_LONG_TERM';
+
+import { OFFER, getStandardProductOffer } from '../../../utils/server';
 
 export default offerType => {
-    switch (offerType) {
-        case 'GPL':
-        case 'GPLQ':
-        case 'GPLNQ':
-        case 'GPLNQ_RANGE':
-            return gpl;
+    switch (getStandardProductOffer(offerType)) {
+        case OFFER.PAYPAL_CREDIT_NO_INTEREST:
+        case OFFER.PAYPAL_CREDIT_INSTALLMENTS:
+            return paypalCredit;
+        case OFFER.PAY_LATER_SHORT_TERM:
+            return payLaterShortTerm;
+        case OFFER.PAY_LATER_LONG_TERM:
         default:
-            return ppc;
+            // Generic message is included with the long term mutations.
+            // Default case catches GENERIC where getStandardProductOffer returns `undefined`
+            return payLaterLongTerm;
     }
 };