+++ tests/functional/v2/spec/snapshot-tests/standalone/multiProduct.test.js
@@ -45,7 +45,7 @@ descFn.each(filterPermutations([LOCALE_CONFIG], [ACCOUNT]))(
         });
 
         test(`Amount:${amount} - Product list tiles send user to correct view - ${viewport}`, async () => {
-            await clickProductListTiles(modalFrame, modalContent);
+            await clickProductListTiles(modalFrame, modalContent, account);
         });
 
         test(`Amount:${amount} - Amount persists between views - ${viewport}`, async () => {