+++ utils/devServerProxy/config/devAccountsV2.config.js
@@ -4,6 +4,7 @@ import AU from './AU';
 import ES from './ES';
 import IT from './IT';
 import FR from './FR';
+import DE from './DE';
 
 // Note that these values are solely mock structures for development purposes
 // that are not guaranteed to be reflective of what is in production
@@ -13,5 +14,6 @@ export default {
     ...AU,
     ...ES,
     ...IT,
-    ...FR
+    ...FR,
+    ...DE
 };